feat: Detail Page 두 덩어리로 분리
- Suspense와 Error Boundary 적용 위함
- edit 모드의 경우 URL Parameter에 mode=edit 추가
